feat/refactor: SeatGuest is working and rename.
